+++ packages/opentelemetry-exporter-collector/src/platform/browser/CollectorExporter.ts
